+++ lib/beamer/locations/habits_location.dart
@@ -1,12 +1,8 @@
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:lotti/beamer/beamer_delegates.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
-import 'package:lotti/pages/create/complete_habit_dialog.dart';
 import 'package:lotti/pages/habits/habits_page.dart';
-import 'package:lotti/themes/theme.dart';
-import 'package:lotti/utils/uuid.dart';
 
 class HabitsLocation extends BeamLocation<BeamState> {
   HabitsLocation(RouteInformation super.routeInformation);
@@ -14,13 +10,10 @@ class HabitsLocation extends BeamLocation<BeamState> {
   @override
   List<String> get pathPatterns => [
         '/habits',
-        '/habits/complete/:habitId',
       ];
 
   @override
   List<BeamPage> buildPages(BuildContext context, BeamState state) {
-    final habitId = state.pathParameters['habitId'];
-
     final pages = [
       BeamPage(
         key: const ValueKey('habits'),
@@ -31,31 +24,6 @@ class HabitsLocation extends BeamLocation<BeamState> {
           child: const HabitsTabPage(),
         ),
       ),
-      if (habitId != null && isUuid(habitId))
-        BeamPage(
-          routeBuilder: (
-            BuildContext context,
-            RouteSettings settings,
-            Widget child,
-          ) {
-            return DialogRoute<void>(
-              context: context,
-              builder: (context) => child,
-              settings: settings,
-              barrierColor: styleConfig().negspace.withOpacity(0.54),
-            );
-          },
-          key: ValueKey('habits-complete-$habitId'),
-          child: HabitDialog(
-            habitId: habitId,
-            beamerDelegate: habitsBeamerDelegate,
-            data: data.toString(),
-          ),
-          onPopPage: (context, delegate, _, page) {
-            habitsBeamerDelegate.beamBack();
-            return false;
-          },
-        ),
     ];
 
     return pages;