+++ lib/pages/dashboards/dashboard_page.dart
@@ -1,23 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:intersperse/intersperse.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/empty_scaffold.dart';
 import 'package:lotti/services/nav_service.dart';
-import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
-import 'package:lotti/widgets/charts/dashboard_health_chart.dart';
-import 'package:lotti/widgets/charts/dashboard_measurables_chart.dart';
-import 'package:lotti/widgets/charts/dashboard_survey_chart.dart';
-import 'package:lotti/widgets/charts/dashboard_workout_chart.dart';
 import 'package:lotti/widgets/charts/empty_dashboards_widget.dart';
-import 'package:lotti/widgets/charts/habits/dashboard_habits_chart.dart';
-import 'package:lotti/widgets/charts/stories/dashboard_story_chart.dart';
-import 'package:lotti/widgets/charts/stories/wildcard_story_chart.dart';
 import 'package:lotti/widgets/charts/utils.dart';
+import 'package:lotti/widgets/dashboards/dashboard_widget.dart';
 import 'package:lotti/widgets/misc/timespan_segmented_control.dart';
 
 class DashboardPage extends StatefulWidget {
@@ -133,7 +125,6 @@ class _DashboardPageState extends State<DashboardPage> {
                   ),
                   const SizedBox(height: 15),
                   DashboardWidget(
-                    dashboard: dashboard,
                     rangeStart: rangeStart,
                     rangeEnd: rangeEnd,
                     dashboardId: widget.dashboardId,
@@ -143,128 +134,6 @@ class _DashboardPageState extends State<DashboardPage> {
             ),
           );
         },
-      ),
-    );
-  }
-}
-
-class DashboardWidget extends StatelessWidget {
-  const DashboardWidget({
-    required this.dashboard,
-    required this.rangeStart,
-    required this.rangeEnd,
-    required this.dashboardId,
-    super.key,
-    this.showTitle = false,
-  });
-
-  final DashboardDefinition dashboard;
-  final DateTime rangeStart;
-  final DateTime rangeEnd;
-  final String dashboardId;
-  final bool showTitle;
-
-  @override
-  Widget build(BuildContext context) {
-    final items = dashboard.items.map((DashboardItem dashboardItem) {
-      return dashboardItem.map(
-        measurement: (DashboardMeasurementItem measurement) {
-          return DashboardMeasurablesChart(
-            measurableDataTypeId: measurement.id,
-            dashboardId: dashboardId,
-            aggregationType: measurement.aggregationType,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-            enableCreate: true,
-          );
-        },
-        healthChart: (DashboardHealthItem healthChart) {
-          return DashboardHealthChart(
-            chartConfig: healthChart,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-          );
-        },
-        workoutChart: (DashboardWorkoutItem workoutChart) {
-          return DashboardWorkoutChart(
-            chartConfig: workoutChart,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-          );
-        },
-        storyTimeChart: (DashboardStoryTimeItem storyChart) {
-          return DashboardStoryChart(
-            chartConfig: storyChart,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-          );
-        },
-        wildcardStoryTimeChart: (WildcardStoryTimeItem storyChart) {
-          return Column(
-            children: [
-              WildcardStoryChart(
-                chartConfig: storyChart,
-                rangeStart: rangeStart,
-                rangeEnd: rangeEnd,
-              ),
-              const SizedBox(height: 10),
-              WildcardStoryWeeklyChart(
-                chartConfig: storyChart,
-                rangeStart: rangeStart,
-                rangeEnd: rangeEnd,
-              ),
-            ],
-          );
-        },
-        surveyChart: (DashboardSurveyItem surveyChart) {
-          return DashboardSurveyChart(
-            chartConfig: surveyChart,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-          );
-        },
-        habitChart: (DashboardHabitItem habitItem) {
-          return DashboardHabitsChart(
-            habitId: habitItem.habitId,
-            dashboardId: dashboardId,
-            rangeStart: rangeStart,
-            rangeEnd: rangeEnd,
-          );
-        },
-      );
-    });
-
-    return Padding(
-      padding: const EdgeInsets.only(top: 16),
-      child: Column(
-        children: [
-          if (showTitle)
-            Text(
-              dashboard.name,
-              style: taskTitleStyle(),
-            ),
-          ...intersperse(const SizedBox(height: 16), items),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Flexible(
-                child: Padding(
-                  padding: const EdgeInsets.only(left: 10),
-                  child: Text(
-                    dashboard.description,
-                    style: chartTitleStyle(),
-                  ),
-                ),
-              ),
-              IconButton(
-                icon: const Icon(Icons.dashboard_customize_outlined),
-                color: styleConfig().primaryTextColor,
-                onPressed: () =>
-                    beamToNamed('/settings/dashboards/$dashboardId'),
-              ),
-            ],
-          ),
-        ],
       ),
     );
   }