+++ lib/pages/settings/habits/habit_details_page.dart
@@ -15,6 +15,7 @@ import 'package:lotti/pages/settings/form_text_field.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/habits/habit_category.dart';
+import 'package:lotti/widgets/habits/habit_dashboard.dart';
 import 'package:lotti/widgets/misc/datetime_bottom_sheet.dart';
 import 'package:lotti/widgets/settings/form/form_switch.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
@@ -105,6 +106,8 @@ class HabitDetailsPage extends StatelessWidget {
                             ),
                             inputSpacer,
                             SelectCategoryWidget(),
+                            inputSpacer,
+                            SelectDashboardWidget(),
                             inputSpacer,
                             inputSpacer,
                             DateTimeField(