+++ lib/widgets/habits/habit_completion_card.dart
@@ -6,6 +6,7 @@ import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
+import 'package:lotti/pages/create/complete_habit_dialog.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/themes/themes.dart';
@@ -38,9 +39,15 @@ class _HabitCompletionCardState extends State<HabitCompletionCard> {
   final JournalDb _db = getIt<JournalDb>();
 
   void onTapAdd() {
-    beamToNamed(
-      '/habits/complete/${widget.habitDefinition.id}',
-      data: ymd(DateTime.now()),
+    showModalBottomSheet<void>(
+      context: context,
+      useRootNavigator: true,
+      isScrollControlled: true,
+      constraints:
+          BoxConstraints(maxHeight: MediaQuery.of(context).size.height * 0.95),
+      builder: (BuildContext context) {
+        return HabitDialog(habitId: widget.habitDefinition.id);
+      },
     );
   }