+++ test/pages/create/complete_habit_dialog_test.dart
@@ -68,7 +68,6 @@ void main() {
                 ),
                 child: HabitDialog(
                   habitId: habitFlossing.id,
-                  beamerDelegate: delegate,
                 ),
               ),
             ),