+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Habit completion in modal bottom sheet instead of dialog
+- Measurement dialog without beamer page
+- Habit completion dialog without beamer page
+
+### Added:
+- Select dashboard to display during habit completion
+- Show dashboard associated with habit in habit completion modal bottom sheet
+
+### Fixed:
+- Sort & filter dashboards in habit definition
+
+## [0.8.302] - 2023-04-03
 ### Fixed:
 - Suggested previous values in measurement capture dialog cannot be selected